refactor(frontend/src/features/commands
/lookup/Home.jsx):

move CreateForm.jsx to features/command/create

for modular structure
